+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed exposure to properly support TEXTURE2D_X
 
 ### Changed
+- Refactor PixelCoordToViewDirWS to be VR compatible and to compute it only once per frame
 
 ## [6.6.0-preview] - 2019-04-01